requirements.txt
langchain
python-dotenv
langchain_community
pypdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
ipykernel
validators==0.28.1
openai